Add conceptURI for ancestor columns to facilitate display in applications. Consolidate display column classes
